+++ model/Content.go
@@ -1,6 +1,6 @@
 package model
 
 type Content struct {
-	Id    float64 `json:"id"`
-	Proxy Proxy   `json:"proxy"`
+	Id    int64 `json:"id"`
+	Proxy Proxy `json:"proxy"`
 }
